feat: Move code to TS
