--- .npmrc
@@ -1,1 +1,0 @@
-unsafe-perm = true
